+++ src/app/app.module.ts
@@ -5,12 +5,15 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpClientModule, HttpClient } from '@angular/common/http';
-import { CoreModule } from './core/core.module';
+//import { CoreModule } from './core/core.module';
 import { SharedModule } from './shared/shared.module';
 
 import { AppRoutingModule } from './app-routing.module';
 
-import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
+//TrulyUI
+import { CoreModule, InputModule, ButtonModule } from 'truly-ui';
+
+
 
 // NG Translate
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
@@ -19,6 +22,7 @@ import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { HomeModule } from './home/home.module';
 
 import { AppComponent } from './app.component';
+import { HeaderComponent } from './header/header.component';
 
 // AoT requires an exported function for factories
 export function HttpLoaderFactory(http: HttpClient): TranslateHttpLoader {
@@ -26,13 +30,13 @@ export function HttpLoaderFactory(http: HttpClient): TranslateHttpLoader {
 }
 
 @NgModule({
-  declarations: [AppComponent],
+  declarations: [AppComponent, HeaderComponent],
   imports: [
-    NgbModule,
+    CoreModule.forRoot({theme: 'default'}),
     BrowserModule,
     FormsModule,
     HttpClientModule,
-    CoreModule,
+    //CoreModule,
     SharedModule,
     HomeModule,
     AppRoutingModule,
@@ -42,7 +46,11 @@ export function HttpLoaderFactory(http: HttpClient): TranslateHttpLoader {
         useFactory: HttpLoaderFactory,
         deps: [HttpClient]
       }
-    })
+    }),
+
+    //TrulyUI
+    InputModule,
+    ButtonModule
   ],
   providers: [],
   bootstrap: [AppComponent]