+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## 0.0.0 (2020-03-18)
+## <small>0.0.1 (2020-04-06)</small>
 
 * Initial Commit ([732e055](https://github.com/Desmoswal/mttcsp-corp/commit/732e055))
 * Quick fix on readme and changed xenial to bionic in travis config ([386050c](https://github.com/Desmoswal/mttcsp-corp/commit/386050c))
+* Quick fix on test case, added bootstrap ([915dc59](https://github.com/Desmoswal/mttcsp-corp/commit/915dc59))
 * Quick fixes, change from scss to css, testing i18n ([eec2823](https://github.com/Desmoswal/mttcsp-corp/commit/eec2823))